requirements.txt
fastapi
uvicorn
python-multipart
pynacl
cryptography
pycryptodome
crypto
